// Source/SelfService/Web/apis/integrations/generated/models/ServiceAccountAccess.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aigonix.Bridge.M3
 * Bridge API - made for Aigonix Studio
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ServiceAccountAccess = {
    Read: 'Read',
    ReadWrite: 'ReadWrite',
    Admin: 'Admin'
} as const;
export type ServiceAccountAccess = typeof ServiceAccountAccess[keyof typeof ServiceAccountAccess];


export function ServiceAccountAccessFromJSON(json: any): ServiceAccountAccess {
    return ServiceAccountAccessFromJSONTyped(json, false);
}

export function ServiceAccountAccessFromJSONTyped(json: any, ignoreDiscriminator: boolean): ServiceAccountAccess {
    return json as ServiceAccountAccess;
}

export function ServiceAccountAccessToJSON(value?: ServiceAccountAccess | null): any {
    return value as any;
}